// src/types/job.type.ts
export type Job = {
  title: string;
  description: string;
  city: string;
  state: string;
  country: string;
  jobType: string;
  employmentType: string;
  experienceLevel: string;
  yearsOfExperience: string;
  qualification: string;
  responsibilities: string[];
  requirements: string[];
  benefits: string[];
  salaryRangeFrom: number;
  salaryRangeTo: number;
  currency: string;
  deadlineDate: string;
};